move data population to a separate file
